refactor(cli): hoist agent/container to cmd/internal/agentcontainer
