+++ cmd/internal/agentcontainer/container.go
@@ -1,4 +1,4 @@
-package container
+package agentcontainer
 
 import (
 	"github.com/devsy-org/devsy/cmd/flags"
