+++ cmd/internal/agentcontainer/credentials_server.go
@@ -1,4 +1,4 @@
-package container
+package agentcontainer
 
 import (
 	"context"
