+++ cmd/internal/agentcontainer/daemon.go
@@ -1,4 +1,4 @@
-package container
+package agentcontainer
 
 import (
 	"context"
