+++ cmd/internal/agentcontainer/openvscode_async.go
@@ -1,4 +1,4 @@
-package container
+package agentcontainer
 
 import (
 	"encoding/json"
