+++ cmd/internal/agentcontainer/setup_devsy_platform_access.go
@@ -1,4 +1,4 @@
-package container
+package agentcontainer
 
 import (
 	"fmt"
