+++ cmd/internal/agentcontainer/setup_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package container
+package agentcontainer
 
 import (
 	"fmt"
