+++ cmd/internal/agentcontainer/ssh_server.go
@@ -1,4 +1,4 @@
-package container
+package agentcontainer
 
 import (
 	"fmt"
